requirements.txt
requests
biopython
psycopg2-binary
schedule
flask
schedule